% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisp.r
\name{rwispfromcsv}
\alias{rwispfromcsv}
\title{Abstraction for extracting data from a CSV file to run the wispcalc function}
\usage{
rwispfromcsv(file)
}
\arguments{
\item{file}{the name of the file}
}
\value{
A matrix with the alternatives and their global utilities, sorted in descending order of utility.
}
\description{
File requirements: 
 - Separated by comma or semicolon
 - Do not use thousands separator
Example file in https://github.com/dioubernardo/rwisp/blob/main/test.csv
}
