% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisp.r
\name{rwispfromcsv}
\alias{rwispfromcsv}
\title{Abstraction for extracting data from a CSV file to run the wispcalc function}
\usage{
rwispfromcsv(file)
}
\arguments{
\item{file}{the name of the file}
}
\value{
list with 3 matrix, ui = ranking and the global ui, normalizedData = normalized data, utilities = utility values
}
\description{
File requirements: 
 - Separated by comma or semicolon
 - Do not use thousands separator
Example file in https://github.com/dioubernardo/rwisp/blob/main/test.csv
}
