% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisp.r
\name{wispcalc}
\alias{wispcalc}
\title{Integrated Simple Weighted Sum Product Method - WISP}
\usage{
wispcalc(data, alternatives, optimizations, weights)
}
\arguments{
\item{data}{A numeric data matrix, columns are the criteria, rows are the alternatives}

\item{alternatives}{A character vector with the identification of alternatives}

\item{optimizations}{A character vector with definition of minimization or maximization for each criterion, expected 'min' or 'max' only}

\item{weights}{A numeric vector with the criteria weights, the sum of all must be 1}
}
\value{
list with 3 matrix, ui = ranking and the global ui, normalizedData = normalized data, utilities = utility values
}
\description{
Implementation of An Integrated Simple Weighted Sum Product Method - WISP
more information see https://doi.org/10.1109/TEM.2021.3075783
}
